SERP-nokia-nac/requirements.txt
fastapi==0.95.2
uvicorn==0.15.0
network_as_code==1.0.0
httpx>=0.24.1,<0.25.0
python-dotenv==1.0.0
pydantic>=1.10.2,<2.0.0
requests==2.31.0